+++ source/xytheta_transa/xytheta_transa/tasks/manager_based/xytheta_transa/__init__.py
@@ -21,3 +21,13 @@ gym.register(
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
     },
 )
+
+gym.register(
+    id="Xytheta-Transa-Offroad-v0.0",
+    entry_point="isaaclab.envs:ManagerBasedRLEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": f"{__name__}.xytheta_transa_offroad_env_cfg:XythetaTransaOffroadEnvCfg",
+        "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
+    },
+)